backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.1
umap-learn==0.5.7
sentence-transformers==3.4.1
torch==2.6.0
scipy==1.15.1
python-multipart==0.0.20
